Added a redirect activity.
Added fragment callbacks to the Navigation Activity.

Implemented the shares and apps with the mainTVfragment.

Implemented fragment callbacks for Servers.

Implemented the Introduction Activity which will only be launched on first run.

Fixed the dual-log in bug.

Fixed the mobile app crash.

Moved the presenters from mainTVfragment to a diff package

Added logo to the Apps in the mainTVFragment.