Integrated Shares as a HeaderItem and its content as ListRowItem. Removed Apps(Permanently) and settings(temporary)
Moved the isATV() and Preferences code to the util/Preference.java

Implmented ViewStubs.

Fixed the ServerFileActivity crash in the android app.

Moved the isATV to a different class CheckAtv in the util package

Added the settings activity in the MainTVFragment

Implemented Sign-out, connection and sign-out preferences.

Removed fragment callback to improve structure.

Added network calls and made the TV independent of external fragment callbacks.

Added settings and separator.

Implemented the sort algorithm in the MainTVFragment

server select under construction.

Reformatted the code and improved the indentation, spacing etc